routes directory structure refactoring
