+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import router from './router';
+import router from './routes/BrowserRouter';
 import store from './store';
 
 import './assets/tailwind.css';
